Working on spawning policy

+++ traffic-simulation/src/commonMain/kotlin/fr/ciadlab/sim/traffic/Spawner.kt
@@ -4,7 +4,7 @@ import fr.ciadlab.sim.math.geometry.Vector2D
 
 class Spawner<Object>(
     var generation: (()->Object)? = null,
-    var strategy: (()->Unit)? = null,
+    var strategy: ((Double)->Unit)? = null,
     var position: Vector2D = Vector2D(0.0, 0.0),
     var direction: Vector2D = Vector2D(1.0, 0.0),
     val onGeneration: MutableList<(Object) -> Unit> = arrayListOf())
@@ -20,6 +20,6 @@ fun <Object> Spawner<Object>.generation(op: ()->Object) {
     this.generation = op
 }
 
-fun <Object> Spawner<Object>.strategy(op: ()->Unit) {
+fun <Object> Spawner<Object>.strategy(op: (Double)->Unit) {
     this.strategy = op
 }
